sends to server
The infinite loop is for crontab, since crontab has issues with creating too many jobs if one has not finished in time. 
Thus the cronjob must start at boot only once. 
it is this or we have to have a file lock that unlocks the file once the previous instance of it has finished running. 
30 seconds maybe a bit to longer, but I will test it further. 

** Things to change ** 
+ftp is pretty insecure, finding a better solution.
+Arguably such a file may not need to be so secure as its only a simple binary file. 
+The folder on the server that hosts the yml is password protected. 
+May not need credentials for Pi and maybe insecure to list them at all in this script. 
+Naming the yml based on name of the parkinglot.